works

+++ app/src/main/java/com/example/gasstations/domain/usecase/DeleteByServerUseCase.kt
@@ -3,6 +3,6 @@ package com.example.gasstations.domain.usecase
 import com.example.gasstations.domain.repository.Repository
 
 class DeleteByServerUseCase(private val repository: Repository) {
-    suspend fun execute(id: Int) =
+    suspend fun execute(id: Long) =
         repository.delete(id)
 }
